requirements.txt
# Telegram Sender - Main Dependencies

# Telegram API
telethon==1.34.0

# Web Framework (Backend)
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0

# Environment & Config
python-dotenv==1.0.0

# Encryption & Security
cryptography==41.0.8

# Database
sqlite3

# Development
requests==2.31.0 